# src/main/java/org/serratec/trabalhoindividual/controller/ClienteController.java
package org.serratec.trabalhoindividual.controller;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import jakarta.validation.Valid;
import org.serratec.trabalhoindividual.model.cliente.ClienteBuscaId;
import org.serratec.trabalhoindividual.model.cliente.ClienteCriar;
import org.serratec.trabalhoindividual.service.ClienteService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping("/cliente")
public class ClienteController {

    private final ClienteService clienteService;

    public ClienteController(ClienteService clienteService) {
        this.clienteService = clienteService;
    }

    @Operation(summary = "Cadastrar um novo cliente")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "201", description = "Cliente criado com sucesso"),
            @ApiResponse(responseCode = "400", description = "Dados informados inválidos"),
            @ApiResponse(responseCode = "409", description = "Dados já cadastrados")
    })
    @PostMapping
    public ResponseEntity<Void> inserir(@Valid @RequestBody ClienteCriar clienteCriar) {
        clienteService.inserir(clienteCriar);
        return ResponseEntity.status(HttpStatus.CREATED).build();
    }

    @Operation(summary = "Listar clientes ou buscar por nome/cpf")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Lista de clientes retornada"),
            @ApiResponse(responseCode = "400", description = "Dados informados inválidos"),
            @ApiResponse(responseCode = "404", description = "Dados não encontrados")
    })
    @GetMapping
    public ResponseEntity<List<ClienteBuscaId>> buscar(
            @RequestParam(required = false) String cpf,
            @RequestParam(required = false) String nome) {
        return ResponseEntity.ok(clienteService.buscar(cpf, nome));
    }

    @Operation(summary = "Remover um cliente")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "204", description = "Cliente removido com sucesso"),
            @ApiResponse(responseCode = "400", description = "Dados informados inválidos"),
            @ApiResponse(responseCode = "404", description = "Dados não encontrados")
    })
    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deletar(@PathVariable UUID id) {
        clienteService.deletar(id);
        return ResponseEntity.noContent().build();
    }
}